Add CR Updates: Rollback unnecessary changes, fix interfaces
