+++ types/dicomweb-client/index.d.ts
@@ -2,6 +2,15 @@ declare module 'dicomweb-client' {
 
   declare namespace api {
 
+    export interface DICOMwebClientRequestHookMetadata {
+      url: string
+      method: string
+    }
+
+    export interface DICOMwebClientRequestHook { 
+      (request: XMLHttpRequest, metadata: DICOMwebClientRequestHookMetadata): XMLHttpRequest
+    }
+
     export interface DICOMwebClientOptions {
       url: string
       qidoURLPrefix?: string
@@ -10,8 +19,8 @@ declare module 'dicomweb-client' {
       headers?: {
         Authorization?: string
       }
-      requestHooks?: function[],
-      errorInterceptor?: function
+      requestHooks?: DICOMwebClientRequestHook[]
+      errorInterceptor?: (request: DICOMwebClientError) => void
     }
 
     export interface DICOMwebClientError {
